requirements.txt
ultralytics
opencv-python
openai-whisper
moviepy
scipy
pandas
stable-ts